[dag] dto 필드 변경 및 table 추가

+++ dto/air_quality.py
@@ -11,8 +11,8 @@ class AirQualityDTO(BaseModel):
     o3: float
     co: float
     so2: float
-    pm10: int
-    pm25: int
+    pm10: float
+    pm25: float
 
     @validator("measure_datetime")
     def string_to_datetime(cls, value: str) -> datetime.datetime:
